requirements.txt
streamlit
matplotlib
seaborn
pandas
faiss-cpu
ollama